# gateway/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Описание эндпоинтов сервиса анонимизации данных",
        "title": "Anonix Gateway API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/detokenize": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Принимает токен, ищет соответствующий mapping и возвращает исходный plaintext",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tokenizer"
                ],
                "summary": "Детокенизация",
                "parameters": [
                    {
                        "description": "Токен для детокенизации",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.DetokenizeSchema"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.DetokenizeRespSchema"
                        }
                    },
                    "400": {
                        "description": "invalid request body / invalid arguments"
                    },
                    "404": {
                        "description": "token not found / token expired"
                    },
                    "500": {
                        "description": "failed to detokenize / unexpected error"
                    }
                }
            }
        },
...
        "/tokenize": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Принимает plaintext и параметры токенизации, возвращает созданный токен и метаданные.\nРежим \"pseudonymize\" — обратимая операция, mapping сохраняется и его можно детокенизировать.\nРежим \"anonymize\" — необратимая операция, mapping нигде не сохраняется,\nответом является schemas.TokenizeResultSchema, такой токен нельзя детокенизировать.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tokenizer"
                ],
                "summary": "Токенизация",
                "parameters": [
                    {
                        "description": "Данные для токенизации",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.TokenizeSchema"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "mode=anonymize",
                        "schema": {
                            "$ref": "#/definitions/schemas.TokenizeResultSchema"
                        }
                    },
                    "400": {
                        "description": "invalid request body / invalid arguments"
                    },
                    "409": {
                        "description": "token already exists"
                    },
                    "500": {
                        "description": "failed to tokenize / unexpected error"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "schemas.DetokenizeRespSchema": {
            "type": "object",
            "properties": {
                "plaintext": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "schemas.DetokenizeSchema": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
...
        "schemas.TokenizeResultSchema": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string",
                    "example": "fio_7f82a1c3"
                }
            }
        },
        "schemas.TokenizeSchema": {
            "type": "object",
            "properties": {
                "deterministic": {
                    "type": "boolean"
                },
                "kind_id": {
                    "type": "integer"
                },
                "mode": {
                    "description": "\"pseudonymize\" | \"anonymize\"",
                    "type": "string",
                    "example": "pseudonymize"
                },
                "plaintext": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "token_ttl": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}